Strip trailing slashes when generating OpenAPI server URLs
These aren't strictly prohibited, but they're almost certainly not
intended.

This most notably occurs with basePath: /. / is the default value, so
should be equivalent to not setting the basePath (currently it's not)
and is clearly intended to mean 'paths are at the server root', not
'paths should be preceeded with an extra slash'.